Added "daemon" for syncing. Really just a long running loop meant to be managed by an init script/upstart.
The daemon connects to an upstream gerrit and listens for
ref-update events. When such an event is reported, the daemon
schedules a sync for that project only after a configurable delay.

A configurable worker thread pool has been implemented to carry out
the syncing tasks.

Updated sample-config.yaml.
Updated version to 0.1.3