Added ability to handle comment-added events to listen for when the label 'Upstream-Ready' is +1'd.
Added service that sets up a listener for downstream and upstream.
Events are pulled from the streams and handled by a shared worker
thread pool.

Events originating from upstream that are of the ref-updated type
trigger syncing.

Events originating from downstream that are of the comment-added
type can trigger pushing code to upstream.

Refactored configuration and how to create SSH and SSHStream
objects.