gerrit-python-tools now tries to impersonate the downstream author.
Upstream users need to have the same gerrit username as downstream
users. Every upstream user will have to add a shared public key.
Gerrit python tools will have sole access to the private key.

When a change goes upstream ready, gerrit-python-tools will get
the gerrit username of the change owner from the change.owner
attribute of the commentadded event. This username will be used to
construct the upstream git remote. The author attribute is the
author of the comment.